Fixed e2e tests (finally!)
