+++ json-server/applications.js
@@ -1,38 +1,38 @@
 module.exports = {
-  'parent1@example.com': {
-    'event1': [{
+  'parent2@example.com': {
+    'd206004a-b0ce-4267-bf07-133e8113aa1b': [{
       "entity$": "-/cd/applications",
       "id": "6aa9dac9-2c95-409a-a788-c595f70ee216",
       "name": "testKid kidokid",
       "dateOfBirth": "2002-05-06T23:00:00.000Z",
-      "eventId": "event1",
+      "eventId": "d206004a-b0ce-4267-bf07-133e8113aa1b",
       "status": "approved",
-      "userId": "child1",
-      "ticketName": "New Ninjas",
+      "userId": "child2",
+      "ticketName": "Laptop Required",
       "ticketType": "ninja",
-      "sessionId": "6e8d9237-f1a9-4ffa-87ee-35f78e9902f1",
+      "sessionId": "69624aec-e254-4636-b4c6-f623fdb0421b",
       "created": "2017-09-23T11:41:13.092Z",
       "deleted": false,
       "attendance": null,
       "dojoId": "70e868a9-f2b2-4b73-8f83-7e3a79dfa150",
-      "ticketId": "5ea048c8-9e03-475c-996b-ecfd8aeea9b6",
+      "ticketId": "e9552002-c0fb-4fa6-8a06-639fae9e5f2a",
       "notes": "N/A"
     }, {
       "entity$": "-/cd/applications",
       "id": "bbcf2fd3-b1eb-4fa1-90d7-2caefad1a499",
       "name": "Guillaume Feliciano",
       "dateOfBirth": "1991-02-17T00:00:00.000Z",
-      "eventId": "event1",
+      "eventId": "d206004a-b0ce-4267-bf07-133e8113aa1b",
       "status": "approved",
-      "userId": "parent1",
-      "ticketName": "New Ninjas",
-      "ticketType": "ninja",
-      "sessionId": "6e8d9237-f1a9-4ffa-87ee-35f78e9902f1",
+      "userId": "parent2",
+      "ticketName": "Parent",
+      "ticketType": "parent-guardian",
+      "sessionId": "69624aec-e254-4636-b4c6-f623fdb0421b",
       "created": "2017-09-23T11:41:13.096Z",
       "deleted": false,
       "attendance": null,
       "dojoId": "70e868a9-f2b2-4b73-8f83-7e3a79dfa150",
-      "ticketId": "5ea048c8-9e03-475c-996b-ecfd8aeea9b6",
+      "ticketId": "7c6d2cb7-0344-4cfd-8808-c0cfebbbe5af",
       "notes": "N/A"
     }]
   }
