+++ test/e2e/page-objects/dojo-details.js
@@ -151,6 +151,11 @@ const DojoDetailsPage = Object.create(InfoColumn, {
       return $$('.cd-event-list-item__date-timestamp')[index];
     }
   },
+  firstEventDate: {
+    get() {
+      return $('.cd-event-list-item__date-timestamp');
+    }
+  },
   eventTimes: {
     value(index) {
       return $$('.cd-event-list-item__times-timestamp')[index];
